+++ corpus/cases/adv-pkg-main-dist/test/a.test.js
@@ -1,0 +1,2 @@
+import { thing } from '../';
+console.log(thing);