+++ corpus/cases/adv-pkg-main-real/test/a.test.js
@@ -1,0 +1,2 @@
+import { ghost_fn } from '../';
+console.log(ghost_fn);